Widen block ids to 16 bits
Chunks, the mesher's padded copy, worker transfers and every per-block
table now hold 16-bit ids (up to 2048 blocks). Existing blocks keep ids
0..255 and items keep 256..1023, so saves are unchanged; new blocks go
from 1024 up.

+++ js/mesher.js
@@ -17,7 +17,7 @@ const PH = CH + 2;             // padded height (y = -1 .. CH)
 const POS_SCALE = 64;          // vertex position units per block
 const P4 = POS_SCALE / 16;     // per 1/16 of a block
 const VERT_BYTES = 20;
-const padBlocks = new Uint8Array(RSS * PH);
+const padBlocks = new Uint16Array(RSS * PH);
 const skyL = new Uint8Array(RSS * PH);
 const blkL = new Uint8Array(RSS * PH);
 const QCAP = RSS * PH;
